# expression: implement vectorized evaluation for `builtinDateLiteralSig`