PCP #12101 

### What problem does this PR solve? <!--add issue link with summary if exists-->
Implement vectorized evaluation for `builtinDateLiteralSig` at #12101 

### What is changed and how it works?
It gets faster about 24x.

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test
```
> GO111MODULE=on go test -check.f TestVectorizedBuiltinTimeFunc
PASS: builtin_time_vec_test.go:237: testEvaluatorSuite.TestVectorizedBuiltinTimeFunc	0.145s
PASS: builtin_time_vec_generated_test.go:192: testEvaluatorSuite.TestVectorizedBuiltinTimeFuncGenerated	0.099s
OK: 2 passed
PASS
ok  	github.com/pingcap/tidb/expression	0.264s
```
 - Benchmark
```
> go test -v -benchmem -bench=BenchmarkVectorizedBuiltinTimeFunc -run=BenchmarkVectorizedBuiltinTimeFunc -args builtinDateLiteralSig
goos: darwin
goarch: amd64
pkg: github.com/pingcap/tidb/expression
BenchmarkVectorizedBuiltinTimeFuncGenerated-12    	1000000000	         0.0182 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinTimeFunc/builtinDateLiteralSig-VecBuiltinFunc-12         	 1303964	       918 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinTimeFunc/builtinDateLiteralSig-NonVecBuiltinFunc-12      	   51921	     22245 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/pingcap/tidb/expression	3.659s
```